basic_requirements.txt
pandas
pymongo
fastapi
sqlalchemy
python-jose
uvicorn
python-multipart
Jinja2
bcrypt
starlette
PyYAML
pymysql
MySQL-python
python-dotenv
